refactor: orderBy for UsernameSearchResults
