+++ wallet/src/de/schildbach/wallet/data/UsernameSearchResult.kt
@@ -19,6 +19,7 @@ import android.os.Parcelable
 import de.schildbach.wallet.database.entity.DashPayContactRequest
 import de.schildbach.wallet.database.entity.DashPayProfile
 import kotlinx.android.parcel.Parcelize
+import java.util.ArrayList
 
 @Parcelize
 data class UsernameSearchResult(val username: String,
@@ -65,5 +66,23 @@ data class UsernameSearchResult(val username: String,
         REQUEST_SENT,
         REQUEST_RECEIVED,
         CONTACT_ESTABLISHED
+    }
+}
+
+fun ArrayList<UsernameSearchResult>.orderBy(orderBy: UsernameSortOrderBy) {
+    when (orderBy) {
+        UsernameSortOrderBy.DISPLAY_NAME -> this.sortBy {
+            if (it.dashPayProfile.displayName.isNotEmpty())
+                it.dashPayProfile.displayName.lowercase()
+            else it.dashPayProfile.username.lowercase()
+        }
+        UsernameSortOrderBy.USERNAME -> this.sortBy {
+            it.dashPayProfile.username.lowercase()
+        }
+        UsernameSortOrderBy.DATE_ADDED -> this.sortByDescending {
+            it.date
+        }
+        else -> { /* ignore */ }
+        //TODO: sort by last activity or date added
     }
 }
